backend/requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-ninja==1.4.5
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
